api/requirements.txt
attrs==19.1.0
beautifulsoup4==4.8.0
blis==0.2.4
Bottleneck==1.2.1
certifi==2019.6.16
chardet==3.0.4
Click==7.0
cycler==0.10.0
cymem==2.0.2
dataclasses==0.6
fastai==1.0.55
fastprogress==0.1.21
Flask==1.1.1
gunicorn==19.9.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
kiwisolver==1.1.0
MarkupSafe==1.1.1
matplotlib==3.1.1
murmurhash==1.0.2
mysql-connector-python==8.0.16
numexpr==2.7.0
numpy==1.16.4
nvidia-ml-py3==7.352.0
packaging==19.1
pandas==0.25.1
Pillow==6.1.0
plac==0.9.6
preshed==2.0.1
protobuf==3.9.1
pyparsing==2.4.2
python-dateutil==2.8.0
pytz==2019.2
PyYAML==5.1.2
requests==2.22.0
scipy==1.3.1
six==1.12.0
soupsieve==1.9.3
spacy==2.1.8
srsly==0.1.0
thinc==7.0.8
torch==1.1.0
torchvision==0.3.0
tqdm==4.32.2
typing==3.7.4.1
urllib3==1.25.3
wasabi==0.2.2
Werkzeug==0.15.5
git+https://github.com/casics/nostril.git